Replace "getUserMedia" with custom wrapper
The normalization of errors by the "getUserMedia" module should be
covered by current webrtc-adapter releases, so it should not be needed.
Besides that, the NPM package contains a "node_modules" directory with
the code of an old "webrtc-adapter" release, and the bundling tool gets
confused when this directory is present and bundles that old code
instead of the most recent one.

Due to all this, the module is replaced by a custom wrapper that just
calls the given callbacks as needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>