Handle bogus behaviour of "getUserMedia" in Edge
"getUserMedia" should resolve the promise only if all the constraints
are satisfied, but Edge resolves it also if both are requested (audio
and video) but only one of them is satisfied. The callbacks are used to
set the local audio and video as available or not available, so the Edge
behaviour is now handled to return the right information to the
callbacks.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>